requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.0
mysqlclient==2.1.1
python-dotenv==0.18.0
dj-rest-auth==2.1.12
django-allauth==0.51.0
djangorestframework-simplejwt==5.0.0
drf-writable-nested==0.6.4
pandas==1.4.3

# mac M1 only 
PyMySQL==1.0.2